Part1/second_approach/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
psycopg2-binary==2.9.9
redis==5.0.1
kafka-python==2.0.2
python-multipart==0.0.6
pydantic==2.5.0
sqlalchemy==2.0.23
